chore(api): add mypy static type checking, fix 106 real errors it found (#121)
This codebase had ruff (style/lint) but no type checker at all — no
mypy, no pyright. That's a real blind spot: nothing catches wrong
function signatures, None-handling gaps, or mismatched types, and the
local/CI test suite can't catch it either, since DATABASE_URL there
always connects as the Postgres superuser (unrestricted), while
production increasingly does not (see PR #119, same session).

Installed mypy with the pydantic plugin, ran it across the whole `app`
package, and fixed every one of the 106 errors it found — not by
suppressing them, but either by fixing the actual bug or correcting a
stale/wrong type annotation to match what the code actually does.
Several were genuine latent bugs, not just type noise:

- app/api/routes/workers.py: `session.get(...)` can return None; the
  drain-toggle endpoint dereferenced it unconditionally. Now 404s
  instead of crashing on a narrow race between the existence check and
  the row fetch.
- app/orchestration/controller.py: `submit_review_decision()` is typed
  to always return an OutboxEvent, but silently returned None on an
  already-decided gate ("idempotent" comment, no actual idempotent
  value). Its only caller immediately dereferences the result — now
  raises a clear error instead of crashing downstream with an opaque
  AttributeError.
- app/orchestration/claiming.py: the provider-budget check was scoped
  to the *worker's* workspace_id, which is nullable (a worker can serve
  all workspaces, migration 0017's own documented design) — passing
  None into a per-workspace budget lock. Rescoped to the *candidate
  assignment's* workspace_id, which is always set. Flagging separately:
  the claim query's own WHERE clause (`StageAssignment.workspace_id ==
  worker.workspace_id`) likely already zero-matches for a
  None-workspace worker, since stage_assignments.workspace_id is NOT
  NULL — a global/shared worker may never be able to claim anything via
  this path today. Left untouched here (a real fix would change
  claiming semantics and needs its own review); worth a follow-up
  issue.
- app/orchestration/controller.py / strategy.py: a couple of spots
  reused a variable name across incompatible types within one function
  (`audit` rebound from a per-opportunity ResearchAudit to a new
  StrategyAudit; `rows` rebound across elif branches with different row
  shapes). Not a runtime crash today since each use stayed local, but
  exactly the kind of thing that breaks on the next innocent edit.
  Renamed for clarity.
- app/models/config.py, app/models/research.py: SpendCap and
  OpportunityEvidence declared `Mapped[float]` for Numeric() columns,
  which SQLAlchemy actually returns as Decimal — every other call site
  already treated them as Decimal via `Decimal(str(...))`. Corrected
  the annotations to match reality (money fields, so this also closes
  a latent float-precision risk, not just a type mismatch).

Everything else was a stale/inaccurate type hint (routes returning ORM
objects under a `-> XOut` annotation that FastAPI's response_model
coerces correctly at runtime; `.scalars().all()` typed as Sequence
against a declared `list[...]` return) — corrected to reflect actual
behavior with no behavior change.

Added mypy as a required CI step (apps/api's `api` job, after ruff,
before tests) so this class of bug can't merge silently again.

Verified: full API suite green (369 passed, 81% coverage, unchanged
from before this PR) on a fresh migrated DB; ruff check/format clean;
mypy clean (0 errors, 123 source files).

+++ apps/api/app/models/config.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import uuid
+from decimal import Decimal
 
 from sqlalchemy import Enum as SAEnum
 from sqlalchemy import Index, Numeric, Text, text
@@ -50,8 +51,8 @@ class SpendCap(Base, WorkspaceScopedMixin, TimestampMixin, ActorMixin, VersionMi
 
     id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     provider: Mapped[str | None] = mapped_column(Text, nullable=True)
-    daily_cap_usd: Mapped[float] = mapped_column(Numeric(12, 4), nullable=False)
-    monthly_cap_usd: Mapped[float] = mapped_column(Numeric(12, 4), nullable=False)
+    daily_cap_usd: Mapped[Decimal] = mapped_column(Numeric(12, 4), nullable=False)
+    monthly_cap_usd: Mapped[Decimal] = mapped_column(Numeric(12, 4), nullable=False)
 
 
 class ProviderCredential(
